Refactor fitting
remove utils.fitting.py
Add mmg_toolbox.fitting module

Split up functions into sub-files

Add additional tests for imports, all tests pass

Add examples to docs.
