+++ examples/ideal_nexus_file.py
@@ -7,7 +7,7 @@ import datetime
 import numpy as np
 import mmg_toolbox.nexus.nexus_writer as nw
 from mmg_toolbox.diffraction.diffcalc import UB
-from mmg_toolbox.utils.fitting import gauss
+from mmg_toolbox.fitting import gauss
 
 
 # Metadata
