+++ mmg_toolbox/fitting/__init__.py
@@ -1,0 +1,14 @@
+
+from .functions import *
+from .models import *
+from .results import *
+from .fit_functions import *
+from .manager import *
+
+__all__ = [
+    'poisson_errors', 'peak_ratio', 'gen_weights', 'gauss', 'group_adjacent', 'find_peaks', 'find_peaks_str', 'max_index',
+    'modelfit', 'peakfit', 'peak2dfit', 'generate_model', 'generate_model_script', 'multipeakfit',
+    'peak_results', 'peak_results_str', 'peak_results_fit', 'peak_results_plot', 'Peak', 'FitResults',
+    'PEAK_PARS', 'METHODS', 'PEAK_MODELS', 'BACKGROUND_MODELS', 'get_peak_model', 'get_background_model', 'get_default_model',
+    'ScanFitManager'
+]
